cleanup (BaseSkeleton): move to `~/components/base`;
